refactor (Price) : 카운트 이기 때문에 long -> int로 변경

+++ src/main/java/lotto/domain/Price.java
@@ -25,8 +25,8 @@ public class Price {
         return price >= LOTTO_PRICE;
     }
 
-    public long lotteryCount() {
-        return price / LOTTO_PRICE;
+    public int lotteryCount() {
+        return (int) price / LOTTO_PRICE;
     }
 
     public Price calculateWinPrice(final int matchCount) {
